Move as much as possible to private
