+++ LootLockerServerSDK/Source/LootLockerServerSDK/Private/LootLockerServerPersistedState.h
