+++ LootLockerServerSDK/Source/LootLockerServerSDK/Private/LootLockerServerResponse.h
