+++ LootLockerServerSDK/Source/LootLockerServerSDK/Private/LootLockerServerStateData.h
