+++ LootLockerServerSDK/Source/LootLockerServerSDK/Private/ServerAPI/LootLockerServerAuthRequest.h
@@ -4,8 +4,9 @@
 
 #include "CoreMinimal.h"
 #include "LootLockerServerConfig.h"
-#include "LootLockerServerHttpClient.h"
 #include "JsonObjectConverter.h"
+#include "LootLockerServerSDK/Private/LootLockerServerResponse.h"
+#include "LootLockerServerSDK/Private/LootLockerServerHttpClient.h"
 #include "LootLockerServerAuthRequest.generated.h"
 
 //==================================================
